Better support for local libraries
If a py_library is included in a venv and uses the imports attr, it will
be linked into the site-packages directory with the proper path.

Additionally, an always_link attr is added that, when True, will allways
link py_library targets into site-packages even if they don't specify
imports.